fix: 코드 컨벤션 네이밍 적용
