+++ src/main/java/in/koreatech/koin/admin/club/dto/request/AdminClubActiveChangeRequest.java
@@ -9,14 +9,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record AdminDecideClubManagerRequest(
+public record AdminClubActiveChangeRequest(
     @NotNull
-    @Schema(description = "동아리 이름", example = "BCSD", requiredMode = REQUIRED)
-    String clubName,
-
-    @NotNull
-    @Schema(description = "동아리 관리자 승인 여부", example = "false", requiredMode = REQUIRED)
-    Boolean isAccept
+    @Schema(description = "동아리 활성화 여부", example = "false", requiredMode = REQUIRED)
+    Boolean isActive
 ) {
-
+    
 }
