+++ src/main/java/in/koreatech/koin/admin/club/dto/request/AdminClubManagerDecideRequest.java
@@ -9,10 +9,14 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record AdminChangeClubActiveRequest(
+public record AdminClubManagerDecideRequest(
     @NotNull
-    @Schema(description = "동아리 활성화 여부", example = "false", requiredMode = REQUIRED)
-    Boolean isActive
+    @Schema(description = "동아리 이름", example = "BCSD", requiredMode = REQUIRED)
+    String clubName,
+
+    @NotNull
+    @Schema(description = "동아리 관리자 승인 여부", example = "false", requiredMode = REQUIRED)
+    Boolean isAccept
 ) {
-    
+
 }
