+++ src/main/java/in/koreatech/koin/domain/club/dto/request/ClubIntroductionUpdateRequest.java
@@ -9,7 +9,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record UpdateClubIntroductionRequest(
+public record ClubIntroductionUpdateRequest(
     @Schema(description = "동아리 상세 소개", example = "BCSD는 스타트업형 IT 동아리로 KOIN을 운영하고 있습니다", requiredMode = REQUIRED)
     @NotEmpty(message = "동아리 상세 소개를 입력해주세요.")
     String introduction
